moved the useful DOM functions into a separate file myDomModule.js
